requirements.txt
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database
motor==3.3.2
pymongo==4.6.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and LLM
openai==1.3.7

# HTTP client for API calls
httpx==0.25.2

# File upload handling
python-multipart==0.0.6

# Type hints
typing-extensions==4.14.0

# Production server (optional, for production deployment)
gunicorn==21.2.0